TST: Rework after moving
