+++ tests/models/test_energy_balance.py
@@ -12,9 +12,7 @@ import numpy as np
 import pytest
 
 import porepy as pp
-
-
-from porepy.applications.test_utils import models
+from porepy.applications.test_utils import models, well_models
 
 
 class BoundaryConditionLinearPressure(
@@ -234,3 +232,82 @@ def test_unit_conversion(units):
     models.compare_scaled_model_quantities(
         setup, reference_setup, secondary_variables, secondary_units, domain_dimensions
     )
+
+
+class MassAndEnergyWellModel(
+    well_models.OneVerticalWell,
+    models.OrthogonalFractures3d,
+    well_models.BoundaryConditionsWellSetup,
+    well_models.WellPermeability,
+    pp.mass_and_energy_balance.MassAndEnergyBalance,
+):
+    pass
+
+
+def test_energy_conservation():
+    """One central vertical well, one horizontal fracture.
+
+    The well is placed such that the pressure at the fracture is log distributed, up to
+    the effect of the low-permeable matrix.
+
+    """
+    # We want low pressures, to ensure energy is not dominated by -p in fluid part.
+    dt = 1e-4
+    params = {
+        # Set impermeable matrix
+        "material_constants": {
+            "solid": pp.SolidConstants(
+                {
+                    "specific_heat_capacity": 1e0,  # Ensure energy stays in domain.
+                    "well_radius": 0.02,
+                    "residual_aperture": 1.0,  # Ensure high permeability in fracture.
+                    "thermal_conductivity": 1e-6,
+                    "permeability": 1e4,  # Reduce pressure effect
+                    "normal_permeability": 1e4,
+                }
+            ),
+            "fluid": pp.FluidConstants(
+                {
+                    "specific_heat_capacity": 1e0,
+                    "thermal_conductivity": 1e-6,
+                    "normal_thermal_conductivity": 1e-6,
+                }
+            ),
+        },
+        # Use only the horizontal fracture of OrthogonalFractures3d
+        "fracture_indices": [2],
+        "time_manager": pp.TimeManager(schedule=[0, dt], dt_init=dt, constant_dt=True),
+        "grid_type": "cartesian",
+    }
+
+    setup = MassAndEnergyWellModel(params)
+    pp.run_time_dependent_model(setup, params)
+    # Check that the total enthalpy equals the injected one.
+    in_val = 1e7
+    u_expected = in_val * dt
+
+    sds = setup.mdg.subdomains()
+    u = setup.volume_integral(setup.total_internal_energy(sds), sds, 1)
+    h_f = setup.fluid_enthalpy(sds) * setup.porosity(sds)
+    h_s = setup.solid_enthalpy(sds) * (pp.ad.Scalar(1) - setup.porosity(sds))
+    h = setup.volume_integral(h_f + h_s, sds, 1)
+    u_val = np.sum(u.evaluate(setup.equation_system).val)
+    h_val = np.sum(h.evaluate(setup.equation_system).val)
+    assert np.isclose(u_val, u_expected, rtol=1e-3)
+    # u and h should be close with our setup
+    assert np.isclose(u_val, h_val, rtol=1e-3)
+    well_intf = setup.mdg.interfaces(codim=2)
+    well_enthalpy_flux = (
+        setup.well_enthalpy_flux(well_intf).evaluate(setup.equation_system).val
+    )
+    well_fluid_flux = setup.well_flux(well_intf).evaluate(setup.equation_system).val
+    h_well = (
+        setup.fluid_enthalpy(setup.mdg.subdomains(dim=0))
+        .evaluate(setup.equation_system)
+        .val
+    )
+    # The well enthalpy flux should be equal to well enthalpy times well fluid flux
+    assert np.isclose(well_enthalpy_flux, h_well * well_fluid_flux, rtol=1e-10)
+    # Well fluid flux should equal the injected fluid. Minus for convention of interface
+    # fluxes from higher to lower domain.
+    assert np.isclose(well_fluid_flux, -1, rtol=1e-10)
